Password : relocate encryption.ts to src/utils/encription
